+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added PersistAutorun task, PersistStartup task, and PersistWMI task
+- Added Screenshot task, added ScreenshotEvent
+- Added displaying image for ScreenShot events/tasks in Grunt interact view, GruntTasking interact view, and Data view
 
 ### Fixed
 - Fixed hosting files issue